msm: camera: isp: Use single tasklet for dual vfe
Use a single tasklet for dual vfe. With different tasklets
if one of them is not scheduled early enough with respect to
the other tasklet then irq composition fails. By using a single
tasklet irq from both vfe will be handled in the order in which
the hardware generated them.

CRs-Fixed: 2009877
Change-Id: I033cb45afd3162773a181fa4a2dfc912b1bdb109
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>